requirements.txt
Flask>=3.1
Flask-RESTful>=0.3
Flask-SQLAlchemy>=3.1
flask-marshmallow>=1.3
Flask-Migrate>=4.1
marshmallow-sqlalchemy>=1.4
psycopg2-binary
python-dotenv
pytest
pytest-cov
pillow
moviepy
cryptography